+++ include/common.h
@@ -6,7 +6,7 @@
 /*   By: spyun <spyun@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/05/19 09:43:50 by spyun         #+#    #+#                 */
-/*   Updated: 2025/05/23 10:40:00 by spyun         ########   odam.nl         */
+/*   Updated: 2025/05/23 16:37:59 by spyun         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -87,12 +87,13 @@ typedef struct s_game
 /* error.c */
 void			ft_mlx_error(t_game *game);
 
-/* free_utils.c */
+/* free_memory.c */
 void			free_map(t_map *map);
 void			free_game(t_game *game);
 
 /* init.c */
 void			init_window(t_game *game);
+void			init_input_system(t_game *game);
 void			init_game_state(t_game *game);
 
 /* init_components.c */
