+++ src/control/keyboard_input.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        ::::::::            */
-/*   key_handler.c                                      :+:    :+:            */
+/*   keyboard_input.c                                   :+:    :+:            */
 /*                                                     +:+                    */
 /*   By: spyun <spyun@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/05/22 08:52:09 by spyun         #+#    #+#                 */
-/*   Updated: 2025/05/23 15:30:21 by spyun         ########   odam.nl         */
+/*   Updated: 2025/05/23 16:21:29 by spyun         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
